server/api/db_utils.py
import psycopg2
import yaml
import os

def connect(): 
    config = {}
    # config directory is at project root: ../../config/db.yml from server/api
    yml_path = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', 'config', 'db.yml'))

    with open(yml_path, 'r') as file:
        config = yaml.load(file, Loader=yaml.FullLoader)

    return psycopg2.connect(
        dbname=config['database'],
        user=config['user'],
        password=config['password'],
        host=config['host'],
        port=config['port']
    )

def exec_get_one(sql, args={}):
    conn = connect()
    cur = conn.cursor()
    cur.execute(sql, args)
    one = cur.fetchone()
    conn.close()
    return one

def exec_commit(sql, args={}):
    conn = connect()
    cur = conn.cursor()
    result = cur.execute(sql, args)
    conn.commit()
    conn.close()
    return result