requirements.txt
psycopg2-binary==2.9.10
PyYAML==6.0.2